+++ tests/testthat/test-check_collinearity.R
@@ -211,3 +211,10 @@ test_that("check_collinearity, hurdle/zi models w/o zi-formula", {
   )
   expect_equal(out$VIF, c(1.05772, 1.05772, 1.06587, 1.06587), tolerance = 1e-4)
 })
+
+test_that("check_collinearity, invalid data", {
+  skip_if(packageVersion("insight") < "0.19.8.2")
+  dd <- data.frame(y = as.difftime(0:5, units = "days"))
+  m1 <- lm(y ~ 1, data = dd)
+  expect_error(check_collinearity(m1), "Can't extract variance-covariance matrix")
+})