+++ NEWS.md
@@ -1,3 +1,20 @@
+# performance 0.10.10
+
+## New supported models
+
+* Rudimentary support for models of class `serp` from package *serp*.
+
+## General
+
+* Improved error messages for `check_model()` when QQ-plots cannot be created.
+
+## Bug fixes
+
+* Fixed issue in `check_normality()` for t-tests.
+
+* Fixed issue in `check_itemscale()` for data frame inputs, when `factor_index`
+  was not a named vector.
+
 # performance 0.10.9
 
 ## Changes
@@ -12,6 +29,24 @@
 
 * `check_itemscale()` gets a `print_html()` method.
 
+* Clarification in the documentation of the `estimator` argument for
+  `performance_aic()`.
+
+* Improved plots for overdispersion-checks for negative-binomial models from
+  package *glmmTMB* (affects `check_overdispersion()` and `check_mnodel()`).
+
+* Improved detection rates for singularity in `check_singularity()` for models
+  from package *glmmTMB*.
+
+* For model of class `glmmTMB`, deviance residuals are now used in the
+  `check_model()` plot.
+
+* Improved (better to understand) error messages for `check_model()`,
+  `check_collinearity()` and `check_outliers()` for models with non-numeric
+  response variables.
+
+* `r2_kullback()` now gives an informative error for non-supported models.
+
 ## Bug fixes
 
 * Fixed issue in `binned_residuals()` for models with binary outcome, where
@@ -19,6 +54,14 @@
 
 * `performance_score()` should no longer fail for models where scoring rules
   can't be calculated. Instead, an informative message is returned.
+
+* `check_outliers()` now properly accept the `percentage_central` argument when
+  using the `"mcd"` method.
+
+* Fixed edge cases in `check_collinearity()` and `check_outliers()` for models
+  with response variables of classes `Date`, `POSIXct`, `POSIXlt` or `difftime`.
+
+* Fixed issue with `check_model()` for models of package *quantreg*.
 
 # performance 0.10.8